% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1_.R
\name{function1}
\alias{function1}
\title{Function1_Generate Values of X from distribution}
\usage{
function1(n = 20)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Function1_Generate Values of X from distribution
}
