% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2_.R
\name{Function2}
\alias{Function2}
\title{Function2_log first # and add another number}
\usage{
Function2(x, y)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Function2_log first # and add another number
}
