% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3_4_5.R
\name{Function3}
\alias{Function3}
\title{Function3_Sum of two terms}
\usage{
Function3(x, y)
}
\arguments{
\item{x}{(any number)}

\item{y}{(any number)}
}
\value{

}
\description{
Function3_Sum of two terms
}
