% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3_4_5.R
\name{function4}
\alias{function4}
\title{Function4_Subtract two DF or numbers}
\usage{
function4(n, y)
}
\arguments{
\item{n}{(can be a df or number)}

\item{y}{}
}
\value{

}
\description{
Function4_Subtract two DF or numbers
}
