% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3_4_5.R
\name{function5}
\alias{function5}
\title{Function5_Subtraction and multiply function}
\usage{
function5(n, y)
}
\arguments{
\item{n}{(any number)}

\item{y}{(any number)}
}
\value{

}
\description{
Function5_Subtraction and multiply function
}
